requirements.txt
django>=4.1.4
djangorestframework
pyyaml
requests
django-cors-headers
mysqlclient
python-dotenv
django-seed
psycopg2-binary
django-debug-toolbar
